Fixed issue #2, updated BP window (WIP)
When running at the same FPS as your monitor, OpenGL will block due to VSync when updating the screen. This left no time available for the message queue to process and therefore locked up SpecBAS.

Moving the OpenGL calls to the display thread frees up the main thread to process messages, and gives a nice speed boost to sprite creation and management.

WIP - adding breakpoint window to finish off the debug update.