FIxed double "click" sound when using menus
Menus make their own click when the user selects them, and some trigger keypress sequences which also cause a click, causing two sounds to trigger.

Fixed by adding a flag byte to the current keypress data to indicate no click to be played if bit 1 of the flag byte is set.