Fixed issue #17 - Cannot use ; with GOSUB
The peephole optimiser will optimise out GOTO and GOSUB with an integer argument to a much faster cached version that stores the address in compiled code to jump to. This, done too late as statements are added due to the ";" used, causes the statement offset table to become misaligned.

Also, ";" could not be parsed, so that's also fixed.