Crash on autosave fixed (?) Fixed ON ERROR not trapping errors. Fixed crash deleting lines from direct command
Autosave was triggered by a frame counter in the refresh thread; now we do it in the interpreter thread for less crashiness.
ON ERROR <statement> was not triggered because error state was responded to before the periodical checks were done. Moved to the main interpreter routine.
Deleting a line by entering its line number in the direct command window crashed due to a previous fix, a better fix has been applied.