The big menu update
WindowMenu and PopupMenu have been updated to allow acclerator chars in their menu item captions - use & as a prefix, and Alt+char to activate.

Most of the  update is changes to allow these controls to listen to alt-key access.

Because of this, alt+nnn where n is a number 0-9 is now a numpad number only.

Windows can now focus (with workarounds for the editor windows) but don't currently repaint when focus changes.