Tooltips update
Added tooltips to the editor and controls. Hover variables to inspect their values, make a selection and hover it to have it evaluated and displayed as a tip.
CLEAR no longer performs a RESTORE.
Fixed selections in the direct command window causing missing chars.
RND no longer upsets the pRNG sequence when evaluated by a tooltip
Fixed off-by-one pixel in detecting clicks in popupmenu
Fixed crash in WindowMenu where items could be selected with the space key before being marked as selected.