Non-permanent menus, faster refresh
SP_WindowMenu can now be marked as non-permanent, requiring a right mouse-button press to show it inside the parent window.

No longer issue a glFinish() call at the end of a frame in our never-ending quest for more stable framerates.

Fixed a bug that causes a crash when double-clicking on empty space past the end of a line in the editor.

Added a comment that IIF cannot be optimised by inlining its contents, even if those expressions inside it evaluate to a constant.

Fix OUT OF DATA error in Restore in some cases, caused by recent ON RESTORE changes

Allow SP_ListBox headers to resize from the right.