Control sizing, editor bugs
Added the HasSized() virtual method to controls so they can, for example, re-align child controls in response to a change in size.
Fixed the floating scrollbar in the debug panel with this change.
Fixed CTRL+G not jumping to the right place in the listing
Fixed slowdown when resizing debug panel due to massive overuse of recompilation