Renumber and selections
RENUMBER caused a bug where lines with a number but no content would end up with the wrong line number.
splitting a line with RETURN in the middle of the line number caused a cursor misplace
Selecting characters by moving left from the end of a line would cause a glitch in line painting.